[BM-293] :recycle: offset long 타입으로 변경

+++ src/main/java/com/saiko/bidmarket/user/controller/dto/UserBiddingSelectRequest.java
@@ -12,14 +12,14 @@ import lombok.Getter;
 @Getter
 public class UserBiddingSelectRequest {
   @PositiveOrZero
-  private final int offset;
+  private final long offset;
 
   @Positive
   private final int limit;
 
   private final Sort sort;
 
-  public UserBiddingSelectRequest(int offset, int limit, Sort sort) {
+  public UserBiddingSelectRequest(long offset, int limit, Sort sort) {
     this.offset = offset;
     this.limit = limit;
     this.sort = sort != null ? sort : END_DATE_ASC;
